pyproject.toml
[project]
name = "marketshield"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.7",
    "easyocr>=1.7.2",
    "faiss-cpu>=1.12.0",
    "google-genai>=1.32.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pypdf>=6.0.0",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "youtube-transcript-api>=1.2.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.11",
]
